+++ R/radae.R
@@ -47,8 +47,7 @@
 #' ), stringsAsFactors = FALSE
 #' )
 #'
-#' adsl <- radsl(cached = TRUE)
-#' adae <- radae(adsl, lookup_aag = aag)
+#' adae <- radae(ADSL, lookup_aag = aag)
 #'
 #' with(
 #'   adae,