+++ NEWS.md
@@ -1,3 +1,6 @@
+# random.cdisc.data 0.3.8.9000
+Issues with no news:
+
 # random.cdisc.data 0.3.8
 * Four variables `ATC1CD`, `ATC2CD`, `ATC3CD`, `ATC4CD` have been added to ADCM data.
 * ADAETTE data set generation via `radaette()` has been enhanced: